CHROMIUM: config: Enable Intel Virtual Button Driver
Enable the Intel Virtual Button Driver.  There will be a later change
on some devices to replace the Google TBMC device with the Intel
Virtual Buttons device.  The TBMC device is a subset of functionality,
and by switching to the device we can eventually get rid of the
need to support and maintain a separate driver for the TBMC device.

BUG=None
TEST=Build.

Signed-off-by: Matt Delco <delco@chromium.org>
Change-Id: Idd6ab905186159c908a3a5ae70e07fa2c3c4054e
Reviewed-on: https://chromium-review.googlesource.com/1074099
Commit-Ready: Matt Delco <delco@google.com>
Tested-by: Matt Delco <delco@google.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>