UPSTREAM: usb: hub: Per-port setting to use old enumeration scheme
The "old" enumeration scheme is considerably faster (it takes
~244ms instead of ~356ms to get the descriptor).

It is currently only possible to use the old scheme globally
(/sys/module/usbcore/parameters/old_scheme_first), which is not
desirable as the new scheme was introduced to increase compatibility
with more devices.

However, in our case, we care about time-to-active for a specific
USB device (which we make the firmware for), on a specific port
(that is pogo-pin based: not a standard USB port). This new
sysfs option makes it possible to use the old scheme on a single
port only.

Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 25244227158e1502062041365a439a54cb8fe673)

BUG=b:36579698
TEST=echo 1 > /sys/bus/usb/devices/usb1/1-0:1.0/usb1-port2/quirks
     Time to active on staff lowers from 1058 ms to 766 ms

Change-Id: If8aa6f434576de6474086aa758189bc817acdda0
Reviewed-on: https://chromium-review.googlesource.com/1068775
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>