CHROMIUM: PM / devfreq: rk3399: Pass devfreq object to DFI without clobbering drvdata
To pass the devfreq object to the DFI device we currently clobber the
drvdata of the edev with the DMC data in rk3399_dmcfreq_probe(). This
is a huge hack and leads to other hacks, like adding a devfreq device
before it's governor (commit d5dcc34e053c9 ("CHROMIUM: devfreq: rk3399:
Fix crash when dfi irq triggered too early"), which is incompatible
with upstream patches we want to integrate.

From the edev device it is possibe to get the DFI device. Add the new
function rockchip_dfi_set_devfreq(), which receives as parameters the
edev and the devfreq objects. This allows to set the devfreq of the DFI
at the end of DMC probe(), when we are sure that the DMC device was
created successfully.

The check for devfreq can be omitted in the ISR of the hard IRQ, since
the ISR doesn't access the devfreq. It now always triggers the threaded
IRQ which verifies that the devfreq has been set before doing any
processing.

BUG=b:65276497, b:74321682, b:73050145
TEST=build for scarlet and boot
  cat /sys/devices/platform/dmc/devfreq/*/cur_freq
    => frequency varies for varying system loads

Change-Id: I013239e2aa7e70e7ca39b5e6023180da523c09eb
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1069776
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>