CHROMIUM: media: ipu3-imgu: Fix interrupt handler return value
Interrupt handler must not return IRQ_NONE unless it fails to detect
hardware interrupt condition. Since imgu_isr_threaded() is always called
as a result of a detected hardware interrupt condition, make it always
return IRQ_HANDLED.

We already fixed this in CL:601807, but it looks like a part of it was
not included in latest FROMLIST series, which we synced up our code
with. Fix it again...

BUG=b:64324464
BUG=b:72725948
TEST=Test on poppy/soraka and make sure there's no unhandled
     ipu3-imgu IRQ while streaming.

Change-Id: If58750f060771d543566b243c9711b0897e31aec
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1088539
Reviewed-by: Ricky Liang <jcliang@chromium.org>