FROMLIST: clk: qcom: Add clk_rcg2_gfx3d_ops for SDM845
To turn on the gpu_gx_gdsc, there is a hardware requirement to
turn on the root clock (GFX3D RCG) first which would be the turn
on signal for the gdsc along with the SW_COLLAPSE. As per the
current implementation of clk_rcg2_shared_ops, it clears the
root_enable bit in the enable() and set_rate() clock ops. But due
to the above said requirement for GFX3D shared RCG, root_enable bit
would be already set by gdsc driver and rcg2_shared_ops should not clear
the root unless the disable is called.

Add support for the same by reusing the existing clk_rcg2_shared_ops
and deriving "clk_rcg2_gfx3d_ops" clk_ops for GFX3D clock to
take care of the root set/clear requirement.

Signed-off-by: Amit Nischal <anischal@codeaurora.org>
(am from https://patchwork.kernel.org/patch/10450115/)

BUG=b:79946281
TEST=Boot and can see graphics

Change-Id: Ic6a53b7402dedf1c7d26e399e4e39903d576accc
Signed-off-by: Douglas Anderson <dianders@chromium.org>