CHROMIUM: iwl7000: nvm: get num of hw addresses from firmware
With NICs that don't read the NVM directly and instead rely on getting
the relevant data from the firmware, the number of reserved MAC
addresses was not added to the API. This caused the driver to assume
there is only one address which results in all interfaces getting the
same address. Update the API to fix this.

While at it, fix-up the comments with firmware api names to actually
match what we have in the firmware.

Change-Id: I00e56f65a2b2490db6ee2c41d2103bde7e7ab643
Signed-off-by: Naftali Goldstein <naftali.goldstein@intel.com>
iwl7000-tree: 3e2f9bfe9701c5173523301690c8a82c018dd283