CHROMIUM: iwl7000: mvm: pull some he_phy_data decoding into a separate function
Pull some of the decoding of he_phy_data into a separate function so
we don't need to check over and over again if it's valid.

While at it, fix the UL/DL bit reporting to be for all but trigger-
based frames.

Change-Id: Ib6e96da9a82a3b39af0cd1ede2d39ce2a91dc97f
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 9316a712177df5309fc35f77b775a809f8943dcc