CHROMIUM: iwl7000: mvm: show more HE radiotap data for TB PPDUs
For trigger-based PPDUs, most values aren't part of the HE-SIG-A
because they're preconfigured by the trigger frame. However, we
still have this information since we used the trigger frame to
configure the hardware, so we can (and do) read it back out and
can thus show it in radiotap.

Change-Id: I10929451366d18a3a17381831e534738a1d3f9f3
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: b9d4da456d3cc5bee7bd254f8b12f72ef3104cb8