UPSTREAM: ath10k: add memory dump support for QCA6174/QCA9377
Add memory dump to the firmware crash data file which is provided to user space
via devcoredump interface. This makes it easier for firmware engineers to debug
firmware crashes.

Due to increased memory consumption the memory dump is disabled by default. To
enable it make sure that bit 3 is set in coredump_mask module parameter:

modprobe ath10k_core coredump_mask=0xffffffff

When RAMDUMP is enabled a buffer for the dump is allocated with vmalloc during
device probe. The actual memory layout is different in hardware versions and
the layouts are defined in coredump.c. The memory is split to regions and, to
get even finegrained control of what to copy, the region can split to smaller
sections as not all registers are readable (which could cause the whole system
to stall).

Signed-off-by: Alan Liu <alanliu@qca.qualcomm.com>
[kvalo@qca.qualcomm.com: refactoring and cleanup]
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>
(cherry picked from commit 703f261dd77f3afb8058a927ca2f4651691f0495)

BUG=b:110227638
TEST=Do basic test

Signed-off-by: Wen Gong <wgong@codeaurora.org>
Change-Id: Icfad5d09fe619832adb263b8daff680a22652fcb
Reviewed-on: https://chromium-review.googlesource.com/1116531
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Wen Gong <wgong@qti-qualcomm.corp-partner.google.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>