CHROMIUM: iio: cros_ec: Add back the legacy timestamp processing code
Turns out the new code is not happy if used with the old firmware.
"New code" consists of:
* CHROMIUM: iio: cros_ec: Spread timestamps better
* CHROMIUM: iio: cros_ec: Add predictive filter on fifo_timestamp

We'll have to keep the old code since there's a ton of corner cases
for the old firmware behavior that's required to pass CTS. The new code
can be enabled based on the tight_timestamps variable.

Ideally we want the new/old kernel behavior based on a feature flag
that the EC exposes (TODO), but for now we allow it to be switched via
the "tight_timestamps" userspace attribute in /sys/bus/iio/devices.

Summary of changes:
* Moved new spreading code from inline to cros_ec_ring_spread_add()
* Copied old spreading code from before commit 3af8b8cf0907
  ("CHROMIUM: iio: cros_ec: Spread timestamps better") to
  cros_ec_ring_spread_add_legacy()
* Added state->tight_timestamps variable, default false
* Added IIO_DEVICE_ATTR for tight_timestamps
* Based on if tight_timestamps==1
    * Conditionally used timestamp filtering
    * Conditionally used cros_ec_ring_spread_add{,legacy}()
* last_batch_len[] is now an int, no need for it to be 64 bits.
* div64 in places
* Small fixups

CQ-DEPEND=CL:1128257
TEST=With old scarlet firmware, CTS still passes
TEST=With new scarlet firmware (and udev hack, CL:1128257), timestamps are
up to the 100us jitter spec
BUG=b/111079027, b/109786990

Change-Id: I6e8b4d53f7172fa47ba8e9725387ec4b0ddbc374
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1123651