FROMLIST: misc: throttler: Add core support for non-thermal throttling
The purpose of the throttler is to provide support for non-thermal
throttling. Throttling is triggered by external event, e.g. the
detection of a high battery discharge current, close to the OCP limit
of the battery. The throttler is only in charge of the throttling, not
the monitoring, which is done by another (possibly platform specific)
driver.

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
(am from https://patchwork.kernel.org/patch/10505615/)

BUG=b:74321682, b:73050145
TEST=build and boot on dru

Conflicts:
  MAINTAINERS
  drivers/misc/Kconfig
  drivers/misc/Makefile

    trivial conflicts caused by upstream adding new drivers and
    MAINTAINER entries

Change-Id: I21336916c3fd4062e8add2f867afc3fcff29f99c
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1102142
Reviewed-by: Brian Norris <briannorris@chromium.org>