CHROMIUM: i2c-qup - add retries for NACK
Whirlwind/Arkham TPM (Infineon SLB9645) is failing during attestation
(part of test provisioning) with this sequence:
 ERR kernel: [   47.104234] i2c_qup 12460000.i2c: HW not capable of reads over 256 bytes
 ERR kernel: [   47.120077] i2c_qup 12460000.i2c: NACK from 20
 ERR kernel: [   47.120530] i2c_qup 12460000.i2c: NACK from 20

In comment #131 (see bug) apronin@ noted the communication may
fail with NACK and retrying up to 3 times could allow the transaction
to succeed. This is entirely speculative since no one has been able
to reproduce this at their desk and verify this will work.

Plan is to monitor CQ and jetstream test lab failures to see if this
reduces TPM failures and revert if it does not.

BUG=b:33758106
TEST=port to 3.14 and boot on whirlwind. Verify TPM is discovered and responds properly.

Change-Id: I394e298f3eae25aca2023d038c7f35a478812b4d
Signed-off-by: Mu Lin <mxt@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1130544
Reviewed-by: Kan Yan <kyan@chromium.org>