UPSTREAM: staging: gasket: remove redundant license information
Now that the SPDX tag is in all gasket files, that identifies the
license in a specific and legally-defined manner.  So the extra GPL text
wording can be removed as it is no longer needed at all.

This is done on a quest to remove the 700+ different ways that files in
the kernel describe the GPL license text.  And there's unneeded stuff
like the address (sometimes incorrect) for the FSF which is never
needed.

BUG=b:79253135
TEST=none
CQ-DEPEND=CL:1153443

Cc: Rob Springer <rspringer@google.com>
Cc: John Joseph <jnjoseph@google.com>
Cc: Ben Chan <benchan@chromium.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit bf9c7a8673831acd67ad4a92b6acd85ffcde174c)
Signed-off-by: Zhongze Hu <frankhu@google.com>

Change-Id: Idf06573e3096e82c7bd47bc8c32d2eb4a1e63b9e
Reviewed-on: https://chromium-review.googlesource.com/1153470
Commit-Ready: Zhongze Hu <frankhu@google.com>
Tested-by: Zhongze Hu <frankhu@google.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>