CHROMIUM: iio: cros_ec: Reset filter when there've been no recent interrupts
When something (ex: CTS) turns off sensors for a while, changes
the state of the CPU (eg: load => heat => clock drift) then starts
the sensors again, the filter is not going to be able to smoothly
transition from the old drift to the new one with the sudden change
in drift.

We should just start from scratch to make sure there's no major
discontinuity in the data. Hopefully there are no big irq jitters
around that time.

For now every time there's a bigger gap than 500ms, we clear the filter
and only start using it again after it has at least 8 irq entries in the
history.

TEST=Change TS_HISTORY_BORED_US to something much lower, now a simple
batch will cause it to clear history, note how there's no big
discontinuity (beyond the irq jitter that the filter is supposed to
solve) in the period graph.
TEST=CTS should fail less, especially around the beginning of a test.
BUG=b:67743747

Change-Id: Ie638451091287d4a7486021d9dff83f394d983c0
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1136000
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>