CHROMIUM: add new syscall whitelist to alt-syscall
This CL adds a new syscall table which can be applied to programs in
CrOS that run in a minijail. Existing syscall tables in this file only
enumerate a subset of syscalls available in Linux, whereas this CL adds
a general-purpose whitelist for CrOS that enumerates every syscall in
Linux for the given architectures. While this generic table can be used
to hook any arbitrary syscall in the kernel, for now we only have plans
to use it to hook the setuid() syscall, as part  of some bookkeeping we
have to do for the feature detailed in this doc:
https://docs.google.com/document/d/1yYpayuOsZ_v-Ey8AZJoUAOVaxpKhojLad7PidCndjRM/edit?pli=1#heading=h.xgjl2srtytjt

Although other syscall tables in this file enumerate syscalls for the
x86-32 architecture, I didn't bother with that here since 32-bit x86
boards are EOL in CrOS and we don't have any need for using this
generic syscall table in any kind of 32-bit compatibility mode.

P.S. coming up with the syscalls for each architecture involved some
gross manual awk commands operating on the source headers in these
files (suggestions welcome if anyone knows a better way of doing this):
https://cs.corp.google.com/chromeos_public/src/third_party/kernel/v3.14/arch/x86/syscalls/syscall_64.tbl
https://cs.corp.google.com/chromeos_public/src/third_party/kernel/v3.14/arch/arm/kernel/calls.S
https://cs.corp.google.com/chromeos_public/src/third_party/kernel/v3.14/include/uapi/asm-generic/unistd.h

BUG=chromium:845640, chromium:649417
TEST=used in conjunction with minijail to hook the setuid() syscall and
change its behavior.
Change-Id: Ia4df89bfe68231eafe8d5fc8baa73dff3c9509d1
Signed-off-by: Micah Morton <mortonm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1138403
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>