WIP: arm64: dts: qcom: sdm845: Add qspi (quad SPI) node
This adds the Quad SPI controller to the main sdm845 device tree file.
Boards will be expected to assign the proper pinctrl depending on how
many chip selects they have hooked up and how many data lines.

This depends on a patch to add the clk ID defines and also the
bindings patch of the Quad SPI.

BUG=b:78307062
TEST=flashrom -p host and compare with what servo reads

Change-Id: Id55ff5b11174a44b709da60533d1a34168d836dd
Signed-off-by: Douglas Anderson <dianders@chromium.org>