UPSTREAM: HID: intel-ish-hid: Prevent loading of driver on Mehlow
On Mehlow Xeon-E workstation, ISH PCI device is enabled but without ISH
firmware. Here the ISH device PCI device id was reused for some non Linux
storage drivers. So this was not done for enabling ISH. But this has a
undesirable side effect for Linux.

Here the ISH driver will be loaded via PCI enumeration and will try to do
reset sequence. But reset sequence will wait till timeout as there is no
real ISH firmware is present to take action. This delay will add to boot
time of Linux (This platform will still continue to boot after this
timeout).

To avoid this boot delay we need to prevent loading of ISH drivers on
this platform. So we need to have hack to avoid treating this device as
ISH on this platform. To identify this workstation, we need some runtime
method. Luckily there are special PCI id on this workstation to
distinguish from the client version of this platform. On client version,
the ISH is supported using same PCI device id. So this change look for
the presence of PCI device IDs A309 and A30A and exit.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Reviewed-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit a1e9a9c0dfe47a3f13734be489b7953a2f8bdd83)
Signed-off-by: Rushikesh S Kadam <rushikesh.s.kadam@intel.com>

BUG=b:109857156
TEST=Build kernel, boot & verify that Intel ISH driver is loaded
Verify for Trackpad input functionality (requires ISH FW to be
loaded)

Change-Id: I53ed8f53077c4c67eaf3431db58bea9f384b412c
Reviewed-on: https://chromium-review.googlesource.com/1307044
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Rushikesh S Kadam <rushikesh.s.kadam@intel.com>
Reviewed-by: Rushikesh S Kadam <rushikesh.s.kadam@intel.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>