INTERNAL FIX: msm: ipa: Comment check of GSI_CMD_TIMEOUT
A kernel build with clang fails with the following error:

  drivers/net/ipa/gsi.o: In function `command':
  drivers/net/ipa/gsi.c:1185: undefined reference to
    `__compiletime_assert_1185'
    drivers/net/ipa/gsi.c:1185:(.text+0x3c8c): relocation
    truncated to fit: R_AARCH64_CALL26 against undefined
    symbol `__compiletime_assert_1185'

The offending statement line is:

BUILD_BUG_ON(GSI_CMD_TIMEOUT > (unsigned long)U32_MAX);

GSI_CMD_TIMEOUT is defined as:

static __always_inline unsigned long msecs_to_jiffies(const unsigned int m)
{
	if (__builtin_constant_p(m)) {
		if ((int)m < 0)
			return MAX_JIFFY_OFFSET;
		return _msecs_to_jiffies(m);
	} else {
		return __msecs_to_jiffies(m);
	}
}

The problem is that clang doesn't see 'm' as constant, since it is a
function parameter (see https://crbug.com/829143). As a result it
calls the external function __msecs_to_jiffies() instead of the inline
function _msecs_to_jiffies(), which means it can't determine the
const-ness of the GSI_CMD_TIMEOUT at compile time, which is needed
for the BUILD_BUG_ON statement.

While it seems preferable how gcc handles the constant propagation it
is my understanding that this is an implementation detail, and clang
isn't really to blame here.

Comment the check to be able to build the driver with clang.

Change-Id: Ie2b6aa5aef40e6753a91dc86eac726c5fc0b4e32
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>