WIP: spi: spi-geni-qcom: Properly control chip select
The SPI driver wasn't properly setting the chip select when a SPI
client tried to change the CS with a zero-byte "cs_change" transfer.
Fix this by implementing set_cs.

Apparently:
* The SPI framework's auto_runtime_pm doesn't doesn't think we need PM
  runtime enabled to call set_cs, so we have to do it manually.
  Probably should fix the SPI framework but for now do it here.
* We actually need to send a command to geni and wait for an interrupt
  to get the CS to change (!), so we have to modify the interrupt
  routine to handle this.

Marking this WIP for now, but it could still probably be squashed into
a newer version of the main geni patch.

Change-Id: I2bda61d5d08a4e01202cd3ee79a1c0c7d53d7a12
Signed-off-by: Dilip Kota <dkota@codeaurora.org>
[dianders: reworked a bit]
Signed-off-by: Douglas Anderson <dianders@chromium.org>