FIXUP: CHROMIUM: intel_idle: dump debug info when S0ix fails
telem_soc_states_display() is selected by INTEL_TELEMETRY, which
can be built as module. It is called from intel_idle.c which has
to be built into the kernel. This means that INTEL_IDLE now has
a hard dependency on INTEL_TELEMETRY, which must _not_ be built
as module.

BUG=b:78841592
TEST=Build x86_64:allmodconfig

Change-Id: I4914c0c88ed24164d13cbf5d89383dd24c9fab13
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1153213
Reviewed-by: Furquan Shaikh <furquan@chromium.org>