BACKPORT: ASoC: da7219: Add delays to capture path to remove DC offset noise
On some platforms it has been noted that a pop noise can be
witnessed when capturing audio, mainly for first time after a
headset jack has been inserted. This is due to a DC offset in the
Mic PGA and so to avoid this delays are required when powering
up the capture path.

This commit rectifies the problem by adding delays post Mic PGA and
post Mixin PGA. The post Mic PGA delay is determined based on
Mic Bias voltage, and is only applied the first time after a
headset jack is inserted.

Signed-off-by: Adam Thomson <Adam.Thomson.Opensource@diasemi.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
tag: for-next
(commit 17c81d2f5a59929c73a2a19fd49fe0b068fda76f)

BUG=b:110373632
TEST=Plug-in headset.
     use arecord or app recording file.
     playing rec file, checking no pop hearing at beginning of sound.

Signed-off-by: Mac Chiang <mac.chiang@intel.com>
Change-Id: Id3c8e2deab80dda6bf398a506d7af816b6ad25ec
Reviewed-on: https://chromium-review.googlesource.com/1163539
Tested-by: JeffRJ_Wang <jeffrj_wang@compal.corp-partner.google.com>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>