Revert "BACKPORT: drm/i915: Force ringbuffers to not be at offset 0"
This reverts commit 44ad28428a96b2081828eaf4e83e0fa28c60d5dd.

This change is suspected in causing i915 hangs on SandyBridge systems.
It was backported from a much later kernel which is very different from
what we have in 3.8 and is different from the solution upstream, and
we've run on these systems for the past 6 years without it, so it's
probably not necessary.

BUG=chromium:844256
TEST=modified platform_MemoryPressure which visits mail.com, usatoday.com
  run on parrot sandy bridge systems

Change-Id: I7ad8a93f42eb27589f1fef6a8936266719753a58
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1176549
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>