CHROMIUM: iwl7000: mvm: don't if a pointer is if we can't be unset
We used to have many versions of statistics notification
coming from the firmware. In one of the cleanup patches,
we forgot to clean the code that checks if data->general
is set. Since it is always set, remove the check.

Change-Id: Ie75042afe0683713b724ef6cdfc1d62e5b1440e3
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 0f5259cbd3b695b1224aeb722b30199d4df113f3