BACKPORT/FROMGIT: drm/amd/display: Add support for toggling DFS bypass
[Why]

If the hardware supports DFS bypass it will always be enabled after
creation of the DCCG. DFS bypass should only be enabled when
the current stream consists of a single embedded panel and the
minimum display clock is below the DFS bypass threshold.

[How]

Add a function to the DCCG table that updates the DFS bypass state
when setting the bandwidth. If the DFS bypass state is changed, the
clock needs to be reprogrammed to reflect this before the DPREFCLK
is updated for audio endpoints. The existing display clock value
is used as the target display clock value when reprogramming since the
resulting change will be equal or larger to the current value.

These changes only specifically target dce110 but do offer a framework
for support on other applicable targets.

Signed-off-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Reviewed-by: David Francis <David.Francis@amd.com>
Acked-by: Bhawanpreet Lakha <Bhawanpreet.Lakha@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>

(cherry picked from commit 5a83c93249098df2ee3b0039ec8f4495b959fcd0
    git://anongit.freedesktop.org/drm/drm of origin/drm-next branch)

 Conflicts:
       drivers/gpu/drm/amd/display/dc/dce/dce_clocks.c
       drivers/gpu/drm/amd/display/dc/dce110/dce110_hw_sequencer.c
       drivers/gpu/drm/amd/display/dc/inc/hw/display_clock.h
	[Hang Zhou] Fix rebase conflicts due to missing latest clock voltage update request redesign.

    BUG=b:117659336
    TEST=Build & Boot on Careena

Change-Id: I80a628875480d335d3f58c91d4792d8044c51538
Reviewed-on: https://chromium-review.googlesource.com/1294551
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hang Zhou <hang.zhou@amd.corp-partner.google.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Signed-off-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1358275