HACK: drm/msm/dpu: Free the irq before we free resources
When I booted up under the right timing conditions I'd get a crash at
bootup in dpu_mdss_irq().  It appears that we were freeing resources
but leaving the IRQ on (waiting for devm to presumably free it at some
point in time later?)

For now let's just free it, but fundamentally this driver appears to
be using devm_...() incorrectly.  Either things should use devm and
generally we shouldn't need to call the free function or we shouldn't
use devm.

BUG=b:113072773
TEST=No longer crash at bootup

Change-Id: I7c826c32d8557b58bf26caa695d1100f65c75fbf
Signed-off-by: Douglas Anderson <dianders@chromium.org>