CHROMIUM: iwl7000: mvm: remove dead code
In iwl_mvm_tx_skb_non_sta(), in case of managed interface,
the AP station was supposed to be used for multicast frames
instead of the auxiliary station to avoid frames possibly
sent to an absent P2P GO. However, when moving to DQA mode,
this was broken as no valid queue was assigned. This is fixed
by a recent patch that directs all non-offchannel traffic to
ap station earlier in the TX path. However, the broken, and
now dead code, remained. Remove it.

Change-Id: I236ca6d3120b41a5332903b037dcd61773648a80
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: 20dc5702b60c0eed027346294acc6ee0d0ccce80