UPSTREAM: drm_dp_cec: add note about good MegaChips 2900 CEC support
A big problem with DP CEC-Tunneling-over-AUX is that it is tricky
to find adapters with a chipset that supports this AND where the
manufacturer actually connected the HDMI CEC line to the chipset.

Add a mention of the MegaChips 2900 chipset which seems to support
this feature well.

BUG=b:136019532
TEST=boot grunt, octopus, run login_LoginSuccess, cheets_StartAndroid

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Reviewed-by: Lyude Paul <lyude@redhat.com>
Acked-by: Alex Deucher <alexander.deucher@amd.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20180827075820.41109-3-hverkuil@xs4all.nl
(cherry picked from commit 9bcf6d9868ae95fc7e5eda3dae5200f234ea5623)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I6fa3ad8254265fd56db344e826801ea6799a3ee3