CHROMIUM: media: Move V4L2_CTRL_FLAG_CAN_STORE far from previous flag
The V4L2_CTRL_FLAG_CAN_STORE flag was added by a downstream CL:232581,
which assigned it a bit next to the previous flag. It was then moved one
bit further due to a collision with upstream, which is now the case
again. Move if far a way from previous flag, so the chance of this
happening again is much lower.

Note that even though this flag is exposed to user space, it is not used
by our user space code, so there is no compatibility concern.

BUG=chromium:878124
TEST=emerge-kevin chromeos-kernel-4_4

Change-Id: I0cfbd5edff41a210e3ddf8043e1245e11fd7ebe9
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1195217
Reviewed-by: Pawel Osciak <posciak@chromium.org>