FIXUP: spi: spi-geni-qcom: Add SPI driver support for GENI based QUP (tx_fifo_width bits/bytes)
The name "tx_fifo_width" is sometimes a number of bits and sometimes a
number of bytes.  Let's make it clear which is which.

Also: the code assumes that the TX FIFO width is the same as the RX
FIFO width.  Let's make that explicit with a comment and a name
change.

Similarly "cur_word_len" is confusing about whether it's talking about
bits or bytes.  Let's clarify it too.

It is intended that this patch be squashed into the patch it fixes to
become a new revision of that patch.

BUG=b:80592084
TEST=Use cros_ec a whole bunch

Change-Id: If3e1600182868350b3a8f6845ef2107d24428e9e
Signed-off-by: Douglas Anderson <dianders@chromium.org>