FIXUP: spi: spi-geni-qcom: Add SPI driver support for GENI based QUP (remove)
The spi_geni_remove() code had some calls in it that made no sense.
Specifically it was manually turning resources off and then had a
pm_runtime_put_noidle() with no matching get call.

Running:

  cd /sys/bus/platform/drivers/geni_spi/
  echo 880000.spi > unbind
  sleep 2
  echo 880000.spi > bind

...showed a bunch of yells.  Let's change the code to just do the
pm_runtime_disable().

...oh, but then if you run the above while constantly doing SPI
transactions then you get a hang.  It turns out that we really should
unregister the master before making it so that transfers will fail.
Since there doesn't appear to be a devm_pm_runtime_enable() we'll have
to move away from devm_spi_register_master().

It is intended that this patch be squashed into the patch it fixes to
become a new revision of that patch.

BUG=b:80592084
TEST=Use cros_ec a whole bunch

Change-Id: Ie802cc7a1d328036c483c584559b2de077eb9bae
Signed-off-by: Douglas Anderson <dianders@chromium.org>