FIXUP: spi: spi-geni-qcom: Add SPI driver support for GENI based QUP (spinlock)
Though unlikely, I believe that the timeout code could possibly race
with the ISR.  Let's add a spinlock back in (it was in v1) to protect
against it.  The biggest case where I have seen things like this
needed is with lots of printk spew to a serial console, which can
cause ludicrously long interrupt latencies.

It is intended that this patch be squashed into the patch it fixes to
become a new revision of that patch.

BUG=b:80592084
TEST=Use cros_ec a whole bunch

Change-Id: I15c0b6d6d803c5313b3c8f43d6afb4e2e3226e62
Signed-off-by: Douglas Anderson <dianders@chromium.org>