FIXUP: spi: spi-geni-qcom: Add SPI driver support for GENI based QUP (__func__)
It's discouraged to include __func__ in dev_xxx() printouts.  Those
printouts will already print the device name in the log and that
together with the string should be unique enough to figure out where
the error is coming from The extra __func__ just makes the strings
uglier and the code harder to read.  If __func__ were truly useful in
all printouts then the dev_xxx() macros would already include it.

Remove all the __func__ from printouts in the spi-geni-qcom driver.

It is intended that this patch be squashed into the patch it fixes to
become a new revision of that patch.

BUG=b:80592084
TEST=Use cros_ec a whole bunch

Change-Id: Ide60d2c9375094ca5f49dcfd01c57d751a95bb78
Signed-off-by: Douglas Anderson <dianders@chromium.org>