FIXUP: spi: spi-geni-qcom: Add SPI driver support for GENI based QUP (dark wizardry)
Add a comment explaining that dark wizardry is the reason why the two
"if (m_cmd & SPI_TX_ONLY)" clauses in setup_fifo_xfer() can't be
combined.  If dark wizardry isn't actually the reason why these
clauses can't be combined then perhaps someone can write down the real
reason (or, if there is no real reason, please combine the clauses).

It is intended that this patch be squashed into the patch it fixes to
become a new revision of that patch.

BUG=b:80592084
TEST=Use cros_ec a whole bunch

Change-Id: I4a6456f94cef7b09b8ca40e5c468e510e2832421
Signed-off-by: Douglas Anderson <dianders@chromium.org>