FIXUP: spi: spi-geni-qcom: Add SPI driver support for GENI based QUP (don't init)
In general you shouldn't init a local variable to something unless the
code is relying on it being initted to that value.  Said another way, this is bad:

  int var = 0;

  if (condition)
    var = a;
  else
    var = b;

...specifically we shouldn't init var to 0 when we're just going to
clobber it.  Initting variables like this has a few bad side effects:
- It hinders the compiler's ability to warn you about using a variable
  you haven't initted.
- It makes the reader wonder why you've initted it.

It is intended that this patch be squashed into the patch it fixes to
become a new revision of that patch.

BUG=b:80592084
TEST=Use cros_ec a whole bunch

Change-Id: Idf7db7fc63d5ff5bf379cdcd21cd2739f10d1d28
Signed-off-by: Douglas Anderson <dianders@chromium.org>